Add --environment to create Pinecone index

+++ packages/cli/src/client/pinecone.ts
@@ -10,23 +10,32 @@ import { forEachChunkedAsync } from '../util/array.js';
 import { set } from '../util/storage.js';
 import type { UpsertVectorOptions, VectorDatabase, QueryOptions } from '../types.js';
 
+const getEnvironmentFromUrl = (url: string) => {
+  const m = url.match(/(?<=svc\.)([a-z0-9-]+)(?=\.pinecone)/);
+  if (m) return m[0];
+  return '';
+};
+
 export class Pinecone implements VectorDatabase {
   token: string;
-  url: string;
+  url?: string;
 
   constructor() {
-    if (!PINECONE_URL) throw new Error('Missing PINECONE_URL environment variable');
     if (!PINECONE_API_KEY) throw new Error('Missing PINECONE_API_KEY environment variable');
     this.token = PINECONE_API_KEY;
     this.url = PINECONE_URL;
   }
 
-  async createIndex(name?: string) {
+  async createIndex(name?: string, environment?: string) {
     if (!name) throw new Error('Please provide a name for the index');
+    if (!this.url && !environment)
+      throw new Error('Please use --environment [env] or set PINECONE_URL environment variable');
 
     set('db', 'type', 'pinecone');
 
-    const indices = await pinecone.listIndexes({ url: this.url, token: this.token });
+    const env = environment ?? (this.url ? getEnvironmentFromUrl(this.url) : '');
+
+    const indices = await pinecone.listIndexes({ environment: env, token: this.token });
 
     if (!indices.includes(name)) {
       const body = {
@@ -36,7 +45,7 @@ export class Pinecone implements VectorDatabase {
         podType: PINECONE_POD_TYPE
       };
 
-      await pinecone.createIndex({ url: this.url, token: this.token, body });
+      await pinecone.createIndex({ environment: env, token: this.token, body });
 
       return [
         `Created new index: ${name} (${OPENAI_OUTPUT_DIMENSIONS}/${PINECONE_METRIC}/${PINECONE_POD_TYPE})`,
@@ -46,10 +55,12 @@ export class Pinecone implements VectorDatabase {
   }
 
   async upsertVectors({ namespace, vectors }: UpsertVectorOptions) {
+    const url = this.url;
+    if (!url) throw new Error('Missing PINECONE_URL environment variable');
     let v = 0;
     await forEachChunkedAsync(vectors, PINECONE_UPSERT_VECTOR_LIMIT, async vectors => {
       const body = { vectors, namespace };
-      const response = await pinecone.upsert({ url: this.url, token: this.token, body });
+      const response = await pinecone.upsert({ url, token: this.token, body });
       if (response.upsertedCount !== vectors.length) console.warn('Pinecone response did not match request:', response);
       v += vectors.length;
     });
@@ -57,10 +68,12 @@ export class Pinecone implements VectorDatabase {
   }
 
   async query({ embedding, namespace }: QueryOptions) {
+    if (!this.url) throw new Error('Missing PINECONE_URL environment variable');
     return pinecone.query({ url: this.url, token: this.token, vector: embedding, namespace });
   }
 
   async clearNamespace(namespace: string) {
+    if (!this.url) throw new Error('Missing PINECONE_URL environment variable');
     const body = { deleteAll: true, namespace };
     await pinecone.deleteVector({ url: this.url, token: this.token, body });
   }
